+++ src/main/java/org/visuals/legacy/animatium/mixins/v1/entity/armor_hurt/MixinEquipmentLayerRenderer_DamageTintArmor.java
@@ -29,12 +29,13 @@ import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import com.llamalad7.mixinextras.sugar.Local;
+import com.mojang.blaze3d.textures.GpuTexture;
 import com.moulberry.mixinconstraints.annotations.IfModAbsent;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.entity.layers.EquipmentLayerRenderer;
 import net.minecraft.client.renderer.entity.state.LivingEntityRenderState;
 import net.minecraft.client.renderer.rendertype.RenderType;
 import net.minecraft.client.renderer.rendertype.RenderTypes;
-import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.resources.Identifier;
 import org.objectweb.asm.Opcodes;
@@ -51,27 +52,30 @@ public abstract class MixinEquipmentLayerRenderer_DamageTintArmor {
     @Unique
     private static final String RENDER_LAYERS_TARGET = "renderLayers(Lnet/minecraft/client/resources/model/EquipmentClientInfo$LayerType;Lnet/minecraft/resources/ResourceKey;Lnet/minecraft/client/model/Model;Ljava/lang/Object;Lnet/minecraft/world/item/ItemStack;Lcom/mojang/blaze3d/vertex/PoseStack;Lnet/minecraft/client/renderer/SubmitNodeCollector;ILnet/minecraft/resources/Identifier;II)V";
 
+    @Unique
+    private static final int animatium$DAMAGE_UV = 196608;
+
     @WrapOperation(method = RENDER_LAYERS_TARGET, at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/rendertype/RenderTypes;armorCutoutNoCull(Lnet/minecraft/resources/Identifier;)Lnet/minecraft/client/renderer/rendertype/RenderType;"))
-    private RenderType animatium$renderLayerArmorTint(final Identifier texture, final Operation<RenderType> original) {
-        if (Animatium.isEnabled() && AnimatiumConfig.instance().other.damageTintArmor) {
-            return RenderTypes.entityCutoutZOffset(texture);
+    private <S> RenderType animatium$renderLayerArmorTint(final Identifier texture, final Operation<RenderType> original, @Local(argsOnly = true, ordinal = 0) final S state) {
+        if (this.animatium$isArmorHurt(state) && this.animatium$isVanillaProportions(texture)) {
+            return RenderTypes.entityCutoutNoCullZOffset(texture);
         } else {
             return original.call(texture);
         }
     }
 
     @WrapOperation(method = RENDER_LAYERS_TARGET, at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/Sheets;armorTrimsSheet(Z)Lnet/minecraft/client/renderer/rendertype/RenderType;"))
-    private RenderType animatium$renderLayerArmorTrimTint(final boolean decal, final Operation<RenderType> original, @Local(name = "sprite") final TextureAtlasSprite sprite) {
-        if (Animatium.isEnabled() && AnimatiumConfig.instance().other.damageTintArmor) {
-            return RenderTypes.entityCutoutZOffset(sprite.atlasLocation());
+    private <S> RenderType animatium$renderLayerArmorTrimTint(final boolean decal, final Operation<RenderType> original, @Local(ordinal = 0) final TextureAtlasSprite sprite, @Local(argsOnly = true, ordinal = 0) final S state) {
+        if (this.animatium$isArmorHurt(state) && !decal) {
+            return RenderTypes.entityCutoutNoCullZOffset(sprite.atlasLocation());
         } else {
             return original.call(decal);
         }
     }
 
     @WrapOperation(method = RENDER_LAYERS_TARGET, at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/rendertype/RenderTypes;armorEntityGlint()Lnet/minecraft/client/renderer/rendertype/RenderType;"))
-    private RenderType animatium$useOverlayArmorGlint(final Operation<RenderType> original) {
-        if (Animatium.isEnabled() && AnimatiumConfig.instance().other.damageTintArmor && AnimatiumConfig.instance().other.glintAffectsArmorTint) {
+    private <S> RenderType animatium$useOverlayArmorGlint(final Operation<RenderType> original, @Local(argsOnly = true, ordinal = 0) final S state) {
+        if (this.animatium$isArmorHurt(state) && AnimatiumConfig.instance().other.glintAffectsArmorTint) {
             return AnimatiumRenderTypes.ARMOR_GLINT;
         } else {
             return original.call();
@@ -79,11 +83,24 @@ public abstract class MixinEquipmentLayerRenderer_DamageTintArmor {
     }
 
     @ModifyExpressionValue(method = RENDER_LAYERS_TARGET, at = @At(value = "FIELD", target = "Lnet/minecraft/client/renderer/texture/OverlayTexture;NO_OVERLAY:I", opcode = Opcodes.GETSTATIC))
-    private <S> int animatium$applyOverlayUV(final int original, @Local(argsOnly = true, name = "state") final S entityRenderState) {
-        if (Animatium.isEnabled() && AnimatiumConfig.instance().other.damageTintArmor && entityRenderState instanceof LivingEntityRenderState livingEntityRenderState) {
-            return OverlayTexture.pack(0, OverlayTexture.v(livingEntityRenderState.hasRedOverlay));
-        } else {
-            return original;
-        }
+    private <S> int animatium$applyOverlayUV(final int original, @Local(argsOnly = true, ordinal = 0) final S state) {
+        return this.animatium$isArmorHurt(state) ? animatium$DAMAGE_UV : original;
+    }
+
+    @Unique
+    private <S> boolean animatium$isArmorHurt(final S state) {
+        return Animatium.isEnabled() &&
+                AnimatiumConfig.instance().other.damageTintArmor &&
+                state instanceof LivingEntityRenderState livingEntityRenderState &&
+                livingEntityRenderState.hasRedOverlay;
+    }
+
+    // Patches (https://github.com/Legacy-Visuals-Project/Animatium/issues/76) temporarily
+    // Prevents damage tint on armor if the respective armor texture proportions do not match vanilla.
+    // TODO/Revisit to patch better/nicer as it currently disables the overlay entirely which is not ideal
+    @Unique
+    private boolean animatium$isVanillaProportions(final Identifier location) {
+        final GpuTexture texture = Minecraft.getInstance().getTextureManager().getTexture(location).getTexture();
+        return texture.getWidth(0) == texture.getHeight(0) * 2;
     }
 }